Clean up checkout test files
